+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* (28 January 2015). Restrict username with email id in configured domains.
 * (26 January 2015). Minimized custom code
 * (23 January 2015). Added Network Settings Page
 * (22 January 2015). Initial Release